rename screen to just Diagnostics
