+++ src/navigation/routesNames.js
@@ -12,11 +12,11 @@ const Routes = {
   BACKUP_SHEET: 'BackupSheet',
   CHANGE_WALLET_SHEET: 'ChangeWalletSheet',
   CHANGE_WALLET_SHEET_NAVIGATOR: 'ChangeWalletSheetNavigator',
-  HARDWARE_WALLET_TX_NAVIGATOR: 'HardwareWalletTxNavigator',
   CONFIRM_REQUEST: 'ConfirmRequest',
   CONNECTED_DAPPS: 'ConnectedDapps',
   CURRENCY_SELECT_SCREEN: 'CurrencySelectScreen',
   CUSTOM_GAS_SHEET: 'CustomGasSheet',
+  DIAGNOSTICS_SHEET: 'DiagnosticsSheet',
   DISCOVER_SCREEN: 'DiscoverScreen',
   ENS_ADDITIONAL_RECORDS_SHEET: 'ENSAdditionalRecordsSheet',
   ENS_ASSIGN_RECORDS_SHEET: 'ENSAssignRecordsSheet',
@@ -30,8 +30,9 @@ const Routes = {
   EXPLAIN_SHEET: 'ExplainSheet',
   PORTAL: 'Portal',
   EXTERNAL_LINK_WARNING_SHEET: 'ExternalLinkWarningSheet',
-  IMPORT_SCREEN: 'ImportScreen',
+  HARDWARE_WALLET_TX_NAVIGATOR: 'HardwareWalletTxNavigator',
   IMPORT_OR_WATCH_WALLET_SHEET: 'ImportOrWatchWalletSheet',
+  IMPORT_SCREEN: 'ImportScreen',
   LEARN_WEB_VIEW_SCREEN: 'LearnWebViewScreen',
   MAIN_EXCHANGE_NAVIGATOR: 'MainExchangeNavigator',
   MAIN_EXCHANGE_SCREEN: 'MainExchangeScreen',
@@ -80,7 +81,6 @@ const Routes = {
   TRANSACTION_DETAILS: 'TransactionDetails',
   WALLET_CONNECT_APPROVAL_SHEET: 'WalletConnectApprovalSheet',
   WALLET_CONNECT_REDIRECT_SHEET: 'WalletConnectRedirectSheet',
-  WALLET_DIAGNOSTICS_SHEET: 'WalletDiagnosticsSheet',
   WALLET_NOTIFICATIONS_SETTINGS: 'WalletNotificationsSettings',
   WALLET_SCREEN: 'WalletScreen',
   WELCOME_SCREEN: 'WelcomeScreen',
