+++ src/screens/Diagnostics/constants.ts
