+++ src/screens/Diagnostics/helpers/getCircularReferenceReplacer.ts
